+++ Assets/Novel/Tests/EditMode/Resources/Scenarios/test_unknown_command.rb
@@ -1,0 +1,7 @@
+mystery_effect 'sparkle'
+mystery_ruby 'oya', 'yomi', :first
+if mystery_flag?
+  bg 'bg_after_unknown'
+else
+  bg 'bg_after_unknown'
+end